Changed ChannelUser to use unsigned int IP's Fixed a small bug in the client timer system added a small test of the timer system into gnutest Added xClient::Message( const Channel*, const string& )

+++ include/ChannelUser.h
@@ -105,7 +105,7 @@ public:
 	/**
 	 * Retrieve this channel user's IP.
 	 */
-	inline const long int& getIP() const ;
+	inline const unsigned int& getIP() const ;
 
 	/**
 	 * Retrieve this channel user's integer representation
